# Adding sync.Pool to Compress Middleware