Adding a sync.Pool for the *gzip.Writer reduces the allocations of the Compress middleware in 50% and gives an increase on execution speed of a 85%
This fix #1643

Here are the Benchmarks

Current version
```
PASS
ok      github.com/labstack/echo/v4     0.004s
goos: linux
goarch: amd64
pkg: github.com/labstack/echo/v4/middleware
BenchmarkGzip-2             7473            151518 ns/op          815268 B/op         33 allocs/op
PASS
ok      github.com/labstack/echo/v4/middleware  1.901s
```

Using a sync.Pool
```
PASS
ok      github.com/labstack/echo/v4     0.005s
goos: linux
goarch: amd64
pkg: github.com/labstack/echo/v4/middleware
BenchmarkGzip-2            47617             23736 ns/op            1430 B/op         16 allocs/op
PASS
ok      github.com/labstack/echo/v4/middleware  1.395s
```